add useEffect on dispatch, separated some components
