+++ src/components/Button/Button.jsx
@@ -1,8 +1,3 @@
-import React from "react";
 import { StyledButton } from "./styled";
-
-function Button(props) {
-  return <StyledButton {...props} onClick={props.onClick}></StyledButton>;
-}
 
 export default StyledButton;
