+++ src/components/ComponentsWrapper/ComponentsWrapper.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { StyledWrapper } from "./styled";
+
+function ComponentsWrapper(props) {
+  return <StyledWrapper {...props}>{props.children}</StyledWrapper>;
+}
+
+export default ComponentsWrapper;
