+++ src/components/ComponentsWrapper/index.js
@@ -1,0 +1,3 @@
+import ComponentsWrapper from "./ComponentsWrapper";
+
+export default ComponentsWrapper;
