+++ src/components/PageTitle/PageTitle.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { TitleWrapper, StyledTitle } from "./styled";
+function PageTitle({ children }) {
+  return (
+    <TitleWrapper>
+      <StyledTitle>{children}</StyledTitle>
+    </TitleWrapper>
+  );
+}
+
+export default PageTitle;
