+++ src/components/PageTitle/index.js
@@ -1,0 +1,2 @@
+import PageTitle from "./PageTitle";
+export default PageTitle;
