+++ src/components/Select/constans.js
@@ -1,8 +1,6 @@
-const options = [
+export const options = [
   { id: 1, key: "favorites", value: "My favorites" },
   { id: 2, key: "rare", value: "Rare Stuff" },
   { id: 3, key: "ponies", value: "Cute Ponies" },
   { id: 4, key: "board", value: "Board games" },
 ];
-
-export { options };
